// src/types/patient-portal.ts
export interface PatientLoginForm {
  login_type: 'email' | 'phone' | 'document'
  email?: string
  phone?: string
  document_number?: string
  clinic_code: string
  remember_me: boolean
}